pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "imgshape"
version = "4.2.0"
description = "imgshape v4.2.0 (Bento Intelligence) — Dataset intelligence layer: deterministic fingerprinting, semantic drift, and bento-box UI."
readme = "README.md"
authors = [
  { name = "Stifler", email = "hillaniljppatel@gmail.com" }
]
license = { text = "MIT" }
requires-python = ">=3.8"

# runtime dependencies (keep core lean; heavy libs optional)
dependencies = [
  "Pillow>=9.0.0",
  "numpy>=1.21.0",
  "matplotlib>=3.4.0",
  "scikit-image>=0.19.0",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "pyyaml>=6.0",
  "requests>=2.28.0",
]

[project.optional-dependencies]
torch = [
  "torch>=1.12.0; platform_system != 'Windows' or python_version >= '3.8'",
  "torchvision>=0.13.0",
]
pdf = [
  "weasyprint>=53.0",
  "reportlab>=3.6.0",
  "pyyaml>=6.0",
]
viz = [
  "plotly>=5.20.0",
  "seaborn>=0.12.0",
]
web = [
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "jinja2>=3.1.0",
]
plugins = [
  "importlib-metadata>=6.0",
]
dev = [
  "pytest>=7.0",
  "black>=23.0",
  "flake8>=3.9",
  "pre-commit>=2.20",
  "mypy>=1.0",
  "build>=1.2",
  "twine>=4.0",
]
full = [
  "torch>=1.12.0",
  "torchvision>=0.13.0",
  "weasyprint>=53.0",
  "reportlab>=3.6.0",
  "pyyaml>=6.0",
  "plotly>=5.20.0",
  "seaborn>=0.12.0",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "jinja2>=3.1.0",
]

[project.urls]
Homepage = "https://github.com/STiFLeR7/imgshape"
Source = "https://github.com/STiFLeR7/imgshape"
Issues = "https://github.com/STiFLeR7/imgshape/issues"
Documentation = "https://stifler7.github.io/imgshape/"

[project.scripts]
imgshape = "imgshape.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

# If you have package data (profiles, plugins, assets) ensure they are included
# either via MANIFEST.in or package_data. Example package-data config:
[tool.setuptools.package-data]
"imgshape" = ["profiles/*.yaml", "plugins/*.py", "assets/**"]

[metadata]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering :: Image Recognition",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "Framework :: FastAPI",
  "Development Status :: 5 - Production/Stable"
]
keywords = ["images", "dataset", "fingerprinting", "augmentation", "pytorch", "machine-learning", "atlas", "v4"]
